local/v-sync/vsync_backend/requirements.txt
asgiref==3.7.2
cffi==1.15.1
contourpy==1.0.7
cycler==0.11.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
md5hash==0.0.6
mysqlclient==2.1.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
sqlparse==0.4.4
tzdata==2023.3
